AgenticAI/query_service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pyodbc
langchain
langchain-openai
langchain-community
faiss-cpu
